[Test][Benchmark] fold benchmark_base tests into one file
Merge the workloads_to_params and bench_kernel contract tests into the
existing benchmark_base unit-test file (renamed to match its content).
